Use newly built binutils to link target libraries
This is important when the newly built toolchains link against
different MSVC libraries than the old toolchain used to build
the new one: If LTO is enabled, GCC will build a LTO linker
plugin for the new toolchain, linking against the new toolchain
MSVC libraries. If it then tries to load this plugin into
the old toolchain ld which was linked against a different
MSVC version, ld will crash and the build will fail.

Avoid this by forcing GCC to use the newly built linker,
which will use the same MSVC libraries as the newly built
LTO plugin.