Add option to completely disable LTO in GCC
When using a pre-existing compiler that used a different C runtime
than the current C runtime target, LTO will never work the way
builds are currently structured. This is because binutils (containing
the linker) and gcc are compiled with different C runtimes, and the
linker can't successfully load and execute gcc's LTO plugin in that
case.

Since wanting to create a toolchain that can produce the correct
binaries with the correct C runtime (even though that toolchain
itself partially uses the wrong runtime) is at the very least a
good intermediate step to creating a toolchain the fully uses the
desired C runtime, this should still be supported.

For this reason add an option --no-gcc-lto that disables LTO
support in GCC.