Expat: update to version 2.4.7
All binaries of versions before 2.4.7 on SourceForge were renamed
so that automated scripts (such as this) broke due. The authors of
Expat did this due to a severe security issue that was fixed in
version 2.4.7. Update the script to use that version, which also
allows us to drop a previously required patch.